Refactor output into module
